% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCORE_IRLS.R
\name{CSCORE_IRLS}
\alias{CSCORE_IRLS}
\title{Iteratively reweighted least squares procedure from CS-CORE for estimating and testing cell-type-specific co-expressions using single cell RNA sequencing data}
\usage{
CSCORE_IRLS(X, seq_depth, covar_weight = "regularized")
}
\arguments{
\item{X}{A n by p matrix of UMI counts, where n denotes the number of cells and p denotes the number of genes}

\item{seq_depth}{A length n vector of sequencing depths}

\item{covar_weight}{A character string. If specified to be "regularized", the weight estimates will be regularized across genes}
}
\value{
A list of three p by p matrices: co-expression estimates, p values and test statistics
}
\description{
Iteratively reweighted least squares procedure from CS-CORE for estimating and testing cell-type-specific co-expressions using single cell RNA sequencing data
}
\examples{
# to be added
}
