% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diff_ego_result}
\alias{diff_ego_result}
\title{GO enrichment results from the differential co-expression vignette}
\format{
\subsection{\code{diff_ego_result}}{

A list of \code{enrichResult} objects, one per module.
}
}
\usage{
diff_ego_result
}
\description{
GO enrichment results (from \code{clusterProfiler::enrichGO}) for all WGCNA modules
identified in the differential co-expression analysis of CD14 Monocytes
(COVID-19 vs. healthy controls).
Saved for rendering the vignette without re-running the full analysis.
}
\keyword{datasets}
