% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCORE_IRLS_cpp.R
\name{has_non_integer}
\alias{has_non_integer}
\title{Check for non-integer values in a matrix}
\usage{
has_non_integer(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A logical value: \code{TRUE} if any element is non-integer, otherwise \code{FALSE}.
}
\description{
This function checks whether a numeric matrix contains any non-integer values.
}
\keyword{internal}
