% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_gene_pair}
\alias{ind_gene_pair}
\title{A simulated independent gene pair}
\format{
\subsection{\code{ind_gene_pair}}{

A list with 2 elements:
\describe{
\item{counts}{A 1972 by 2 count matrix with 1972 cells and 2 genes}
\item{seq_depths}{A length 1972 vector of sequencing depths}
}
}
}
\usage{
ind_gene_pair
}
\description{
This data set is created only for illustrative purposes and is used to test CSCORE_IRLS.R.
The source code for generating this data set is in data-raw/ on Github.
}
\keyword{datasets}
