% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_enrich_go}
\alias{top_enrich_go}
\title{Top GO enrichment result in the CS-CORE vignette}
\format{
\subsection{\code{top_enrich_go}}{

A list with five elements, where each element holds the top 3 GO terms enriched in a gene module with strong enrichment signals
}
}
\usage{
top_enrich_go
}
\description{
This data set stores the top GO enrichment result from the analysis in CS-CORE vignette.
It is saved for efficiently rendering the vignette without re-evaluating the codes.
}
\keyword{datasets}
