% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLS_helper.R
\name{WLS_cov}
\alias{WLS_cov}
\title{Weighted Least Squares Covariance Estimation}
\usage{
WLS_cov(D, X, W)
}
\arguments{
\item{D}{Design matrix for WLS (size n x k)}

\item{X}{Gene expression matrix (n x p)}

\item{W}{Weight matrix (n x p)}
}
\value{
A list with:
\describe{
\item{cov_hat}{A k x p x p array of WLS estimates}
\item{test_stat}{A p x p matrix of test statistics}
}
}
\description{
Computes WLS estimates of gene-gene covariance and test statistics that assess
the statistical significance of co-expression using moment-based regressions.
}
