% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLS_helper.R
\name{WLS_mean}
\alias{WLS_mean}
\title{Weighted Least Squares Mean and Variance Estimation (Rcpp)}
\usage{
WLS_mean(D, X, W)
}
\arguments{
\item{D}{Design matrix for WLS (size n x k), where the first column represents baseline mean/var,
and others represents covariates.}

\item{X}{Response matrix for WLS (size n x p)}

\item{W}{Weight matrix (n x p)}
}
\value{
A k x p matrix of WLS estimates
}
\description{
Compute WLS estimates of mean or variance for all p genes
}
