% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCORE_IRLS.R
\name{CSCORE_IRLS}
\alias{CSCORE_IRLS}
\title{Iteratively reweighted least squares (IRLS) procedure in CS-CORE}
\source{
Su, C., Xu, Z., Shan, X., Cai, B., Zhao, H., & Zhang, J. (2023).
Cell-type-specific co-expression inference from single cell RNA-sequencing data.
\emph{Nature Communications}.
doi: \url{https://doi.org/10.1038/s41467-023-40503-7}
}
\usage{
CSCORE_IRLS(
  X,
  seq_depth,
  covariates = NULL,
  post_process = TRUE,
  covariate_level = "z",
  adjust_setting = c(mean = T, var = T, covar = T),
  return_all = FALSE
)
}
\arguments{
\item{X}{A numeric matrix of UMI counts (\code{n x p}),
where \code{n} is the number of cells and \code{p} is the number of genes.}

\item{seq_depth}{A numeric vector of sequencing depths of length \code{n}.}

\item{covariates}{Optional. A numeric matrix of covariates (\code{n x K}) to be adjusted for in the moment-based regressions.
Can be a length n vector if \eqn{K=1}.
If \code{NULL}, no covariates are adjusted for in the regression.
Defaults to \code{NULL}.}

\item{post_process}{Optional. Logical; whether to rescale the estimated co-expressions to lie between –1 and 1.
Defaults to \code{TRUE}.}

\item{covariate_level}{Optional. A character string indicating whether covariates are assumed to affect
the underlying gene expression levels (\code{"z"}) or the observed counts (\code{"x"}).
See the \emph{Details} section for further explanation.
Defaults to \code{"z"}.}

\item{adjust_setting}{Optional. A named logical vector of length 3;
whether to adjust for covariates at the mean, variance, and covariance level.
Must be named \code{c("mean", "var", "covar")}.
Defaults to \code{c(mean = T,var = T, covar = T)}.}

\item{return_all}{Logical; whether to return all estimates, including the effect sizes for covariates.
Defaults to \code{FALSE}.}
}
\value{
A list containing the following components:
\describe{
\item{est}{A \eqn{p \times p} matrix of co-expression estimates.}
\item{p_value}{A \eqn{p \times p} matrix of p-values.}
\item{test_stat}{A \eqn{p \times p} matrix of test statistics evaluating the statistical significance of co-expression.}
\item{mu_beta}{A \eqn{k \times p} matrix of regression coefficients from the mean model. Returned if \code{return_all = TRUE}.}
\item{sigma2_beta}{A \eqn{k \times p} matrix of regression coefficients from the variance model. Returned if \code{return_all = TRUE}.}
\item{cov_beta}{A \eqn{k \times p \times p} array of regression coefficients from the covariance model. Returned if \code{return_all = TRUE}.}
}
}
\description{
This function implements the IRLS procedure used in CS-CORE for estimating and testing
cell-type-specific co-expression from single-cell RNA sequencing data.
}
\details{
Let \eqn{x_{ij}} denote the UMI count of gene \eqn{j} in cell \eqn{i};
\eqn{s_i} denote the sequencing depth;
\eqn{\mu_j,\sigma_{jj}, \sigma_{jj'}} denote the mean, variance and covariance;
\eqn{c_{ik}} denote additional covariate \eqn{k} for cell \eqn{i} (e.g. disease status or cellular state).
The procedure consists of two main steps:
\enumerate{
\item \strong{Mean and variance estimation:} Estimate gene-specific mean and variance parameters
using two moment-based regressions:
\itemize{
\item \strong{Mean model:} \eqn{x_{ij} = s_i (\mu_j + \sum_k c_{ik} \beta_k) + \epsilon_{ij}}
\item \strong{Variance model:} \eqn{(x_{ij} - s_i \mu_{ij})^2 = s_i \mu_{ij} + s_i^2 (\sigma_{jj} + \sum_k c_{ik} \gamma_k) + \eta_{ij}},
where \eqn{\mu_{ij} = \mu_j + \sum_k c_{ik} \beta_k}
}

\item \strong{Covariance estimation and hypothesis testing:} Estimate gene-gene covariance
and compute test statistics to assess the statistical significance of gene co-expression using a third moment-based regression:
\itemize{
\item \strong{Covariance model:} \eqn{(x_{ij} - s_i \mu_{ij})(x_{ij'} - s_i \mu_{ij'}) = s_i^2 (\sigma_{jj'} + \sum_k c_{ik} \theta_k) + \xi_{ijj'}}
}
}
We note that
\enumerate{
\item The formulation above assumes that the covariates alter the mean / variance / covariance of underlying gene expression,
rather than observed counts. If you believe that the covariates directly affect the observed counts independent of underlying gene expression
(e.g. \eqn{x_{ij}|z_{ij} \sim \text{Poisson}(s_i z_{ij}+\sum_k c_{ik} \beta_k)} or \eqn{x_{ij} = s_i \mu_j  + \sum_k c_{ik} \beta_k + \epsilon_{ij}}),
please specify \code{covariate_level="x"}.
\item The original CS-CORE published in \url{https://doi.org/10.1038/s41467-023-40503-7}
did not consider adjusting for covariates \eqn{c_{ik}}'s. This is equivalent to setting \code{covariates} to \code{NULL}.
}
}
\keyword{internal}
