% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCORE_IRLS_base.R
\name{CSCORE_IRLS_base}
\alias{CSCORE_IRLS_base}
\title{Original implementation Iteratively reweighted least squares (IRLS) procedure (archived)}
\source{
Su, C., Xu, Z., Shan, X., Cai, B., Zhao, H., & Zhang, J. (2023).
Cell-type-specific co-expression inference from single cell RNA-sequencing data.
\emph{Nature Communications}.
doi: \url{https://doi.org/10.1038/s41467-023-40503-7}
}
\usage{
CSCORE_IRLS_base(X, seq_depth, post_process = TRUE)
}
\arguments{
\item{X}{A n by p matrix of UMI counts, where n denotes the number of cells and p denotes the number of genes}

\item{seq_depth}{A length n vector of sequencing depths}

\item{post_process}{Whether to process the estimated co-expressions such that the estimates are between -1 and 1. Default to TRUE.}
}
\value{
A list of three p by p matrices:
\describe{
\item{est}{co-expression estimates}
\item{p_value}{p values}
\item{test_stat}{test statistics}
}
}
\description{
This function was originally implemented in 2023 and included in the first release of the CS-CORE R package.
It has since been replaced by \code{\link{CSCORE_IRLS}} for two main reasons:
}
\details{
\enumerate{
\item It relies on base R for regression, which is slower and more memory-intensive than the Rcpp-based implementation in \code{CSCORE_IRLS}.
\item It does not support covariate adjustment, a feature supported in \code{CSCORE_IRLS}.
}
}
\note{
This function is retained for reference and backward compatibility, but users are encouraged to use \code{\link{CSCORE_IRLS}} for new analyses.
}
\examples{
## Toy example:
## run CSCORE on a simulated independent gene pair
cscore_example <- CSCORE_IRLS_base(ind_gene_pair$counts, ind_gene_pair$seq_depths)

## Estimated co-expression between two genes
cscore_example$est[1,2]
# close to 0: 0.007820124

## p-values
cscore_example$p_value[1,2]
# not significant: 0.961981

}
\keyword{internal}
